Update TERRCOST.DAT
There is a bug in EFS V1.5 that causes the AI to crash when setting the movement of Wheel, Tread, & Crawler to 0 due to the AI not correctly checking if it can land in that hex. So for now, the movement costs have been changed to allow these units to move, albeit at significantly higher movement costs.